+++ packages/flow-ui/flow-ui-theme/src/theme/forms.js
@@ -12,6 +12,7 @@ export default {
     }
   },
   input: {
+    fontFamily: `body`,
     border: `none`,
     bg: `omegaLight`,
     color: `omegaDark`,