+++ packages/flow-ui/flow-ui-widgets/src/Sponsor/Sponsor.jsx
@@ -7,8 +7,10 @@ import Section from '@components/Section'
 
 const styles = {
   image: {
-    borderRadius: `default`,
-    overflow: `hidden`
+    overflow: `hidden`,
+    img: {
+      borderRadius: `lg`
+    }
   },
   caption: {
     display: `block`,