+++ starters/gatsby-starter-flexiblog-science/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [3.2.1](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-starter-flexiblog-science@3.2.0...@elegantstack/gatsby-starter-flexiblog-science@3.2.1) (2021-03-30)
+
+
+### Refactor
+
+* **starters:** replace theme version with asterisk ([45db5aa](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/45db5aa9edc36644625c885c4804cd48dc983e62))
+
+
+
+
+## [3.2.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-starter-flexiblog-science@3.1.0...@elegantstack/gatsby-starter-flexiblog-science@3.2.0) (2021-03-27)
+
+
+### Features
+
+* upgrade dependencies ([da85e59](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/da85e59915b171796803e5e281fae0cd2e263e3c))
+
+
+
+
 ## [3.1.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-starter-flexiblog-science@3.0.1...@elegantstack/gatsby-starter-flexiblog-science@3.1.0) (2021-03-20)